pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.12",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain>=0.3.25",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "starlette>=0.46.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.34.2",
    "werkzeug>=3.1.3",
]
